split SystemConfig header and implementation
