+++ system_config/CMakeLists.txt
@@ -28,6 +28,7 @@ include_directories(
 )
 
 add_library(system_config
+  src/SystemConfig.cpp
   src/Configuration.cpp
 )
